refactor: move fancy to its own crate (#1722)
