+++ crates/fancy_display/src/lib.rs
@@ -1,0 +1,3 @@
+pub trait FancyDisplay {
+    fn fancy_display(&self) -> console::StyledObject<&str>;
+}
