+++ crates/pixi_manifest/Cargo.toml
@@ -10,11 +10,13 @@ version = "0.1.0"
 
 [dependencies]
 dunce = { workspace = true }
+fancy_display = { workspace = true }
 indexmap = { workspace = true }
 itertools = { workspace = true }
 lazy_static = { workspace = true }
 pep440_rs = { workspace = true }
 pep508_rs = { workspace = true }
+pixi_consts = { workspace = true }
 regex = { workspace = true }
 serde = { workspace = true }
 serde-untagged = { workspace = true }
@@ -36,11 +38,14 @@ rattler_solve = { workspace = true, features = ["serde"] }
 rattler_virtual_packages = { workspace = true }
 
 # TODO: Remove these dependencies
+console = "0.15.8"
 miette = { workspace = true }
 
 [dev-dependencies]
 assert_matches = { workspace = true }
+fancy_display = { workspace = true }
 insta = { workspace = true, features = ["yaml"] }
+pixi_consts = { workspace = true }
 rstest = { workspace = true }
 serde_json = { workspace = true }
 tempfile = { workspace = true }
