+++ src/cli/install.rs
@@ -1,7 +1,7 @@
 use crate::environment::get_up_to_date_prefix;
-use crate::fancy_display::FancyDisplay;
 use crate::Project;
 use clap::Parser;
+use fancy_display::FancyDisplay;
 use itertools::Itertools;
 use pixi_config::ConfigCli;
 use std::path::PathBuf;
