+++ src/cli/list.rs
@@ -7,10 +7,10 @@ use console::Color;
 use human_bytes::human_bytes;
 use itertools::Itertools;
 
-use crate::fancy_display::FancyDisplay;
 use crate::lock_file::{UpdateLockFileOptions, UvResolutionContext};
 use crate::utils::uv::pypi_options_to_index_locations;
 use crate::Project;
+use fancy_display::FancyDisplay;
 use pixi_manifest::FeaturesExt;
 use pypi_modifiers::pypi_tags::{get_pypi_tags, is_python_record};
 use rattler_conda_types::Platform;
