+++ src/cli/run.rs
@@ -11,7 +11,6 @@ use pixi_config::ConfigCli;
 
 use crate::activation::CurrentEnvVarBehavior;
 use crate::environment::verify_prefix_location_unchanged;
-use crate::fancy_display::FancyDisplay;
 use crate::lock_file::LockFileDerivedData;
 use crate::lock_file::UpdateLockFileOptions;
 use crate::progress::await_in_progress;
@@ -23,6 +22,7 @@ use crate::task::{
     SearchEnvironments, TaskAndEnvironment, TaskGraph,
 };
 use crate::Project;
+use fancy_display::FancyDisplay;
 use pixi_manifest::TaskName;
 use thiserror::Error;
 use tracing::Level;
